token kirjautuminen valmis

+++ osa4/blogilista/controllers/blogs.js
@@ -1,40 +1,67 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs)
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+  response.json(blogs.map(blogs => blogs.toJSON()))
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  if (!blog.title || !blog.url) {
-    response.status(400).json(blog)
-  } else {
-    await blog.save()
-    response.status(201).json(blog)
-  }
-})
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-  const blog = {
-    _id: body._id,
+  const user = request.user
+
+  const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    __v: body.__v
-  }
+    user: user._id
+  })
 
-  await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-  response.status(200).json(blog)
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog.toJSON())
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+blogsRouter.put('/:id', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
+  const blogToUpdate = await Blog.findById(request.params.id)
+
+  const newBlog = {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes
+  }
+
+  if (blogToUpdate.user.toString() === user.id.toString()) {
+    const savedBlog = await Blog.findByIdAndUpdate(request.params.id, newBlog, { new: true })
+    response.json(savedBlog.toJSON())
+  } else {
+    response.status(401).json({ error: 'unauthorized user' })
+  }
+})
+
+
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    response.status(404).json({ error: 'invalid id' })
+  }
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'unauthorized user' })
+  }
 })
 
 module.exports = blogsRouter
